move nametuple to sqlname
