+++ yb-voyager/src/namereg/namereg_test.go
@@ -10,399 +10,70 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/yugabyte/yb-voyager/yb-voyager/src/srcdb"
 	"github.com/yugabyte/yb-voyager/yb-voyager/src/tgtdb"
+	"github.com/yugabyte/yb-voyager/yb-voyager/src/utils/sqlname"
 )
 
-func TestPGDefaultSchemaCaseInsensitiveTableName(t *testing.T) {
-	assert := assert.New(t)
-
-	// Test NewTableName() with PostgreSQL and default schema "public" and
-	// a table name belonging to default schema.
-	tableName := newObjectName(POSTGRESQL, "public", "public", "table1")
-	assert.NotNil(tableName)
-	expectedTableName := &ObjectName{
-		SchemaName:        "public",
-		FromDefaultSchema: true,
-		Qualified: identifier{
-			Quoted:    `public."table1"`,
-			Unquoted:  "public.table1",
-			MinQuoted: "public.table1",
-		},
-		Unqualified: identifier{
-			Quoted:    `"table1"`,
-			Unquoted:  "table1",
-			MinQuoted: "table1",
-		},
-		MinQualified: identifier{
-			Quoted:    `"table1"`,
-			Unquoted:  "table1",
-			MinQuoted: "table1",
-		},
-	}
-	assert.Equal(expectedTableName, tableName)
-	assert.Equal(tableName.MinQualified, tableName.Unqualified)
+var oracleToYBNameRegistry = &NameRegistry{
+	SourceDBType:              ORACLE,
+	role:                      TARGET_DB_IMPORTER_ROLE,
+	SourceDBSchemaNames:       []string{"SAKILA"},
+	YBSchemaNames:             []string{"public"},
+	DefaultSourceDBSchemaName: "SAKILA",
+	DefaultYBSchemaName:       "public",
+	//DefaultSourceReplicaDBSchemaName: "SAKILA_FF", // Will be set using SetDefaultSourceReplicaDBSchemaName().
+	SourceDBTableNames: map[string][]string{
+		"SAKILA": {`TABLE1`, `TABLE2`, `Table2`, `MixedCaps`, `MixedCaps1`, `MixedCAPS1`, `lower_caps`},
+	},
+	YBTableNames: map[string][]string{
+		"public": {"table1", "table2", `Table2`, `mixedcaps`, `MixedCaps1`, `MixedCAPS1`, "lower_caps"},
+	},
 }
 
-func TestPGNonDefaultSchemaCaseInsensitiveTableName(t *testing.T) {
-	assert := assert.New(t)
-	// Test NewTableName() with PostgreSQL and default schema "public" and
-	// a table name belonging to a non-default schema.
-	tableName := newObjectName(POSTGRESQL, "public", "schema1", "table1")
-	assert.NotNil(tableName)
-	expectedTableName := &ObjectName{
-		SchemaName:        "schema1",
-		FromDefaultSchema: false,
-		Qualified: identifier{
-			Quoted:    `schema1."table1"`,
-			Unquoted:  "schema1.table1",
-			MinQuoted: "schema1.table1",
-		},
-		Unqualified: identifier{
-			Quoted:    `"table1"`,
-			Unquoted:  "table1",
-			MinQuoted: "table1",
-		},
-		MinQualified: identifier{
-			Quoted:    `schema1."table1"`,
-			Unquoted:  "schema1.table1",
-			MinQuoted: "schema1.table1",
-		},
+func buildNameTuple(reg *NameRegistry, sourceSchema, sourceTable, targetSchema, targetTable string) *sqlname.NameTuple {
+	var sourceName *sqlname.ObjectName
+	var targetName *sqlname.ObjectName
+	if sourceSchema != "" && sourceTable != "" {
+		sourceName = sqlname.NewObjectName(reg.SourceDBType, sourceSchema, sourceSchema, sourceTable)
 	}
-	assert.Equal(expectedTableName, tableName)
-	assert.Equal(tableName.MinQualified, tableName.Qualified)
-}
-
-func TestPGDefaultSchemaCaseSensitiveTableName(t *testing.T) {
-	assert := assert.New(t)
-	// Test NewTableName() with PostgreSQL and default schema "public" and
-	// a case-sensitive name with mixed cases.
-	tableName := newObjectName(POSTGRESQL, "public", "public", "Table1")
-	assert.NotNil(tableName)
-	expectedTableName := &ObjectName{
-		SchemaName:        "public",
-		FromDefaultSchema: true,
-		Qualified: identifier{
-			Quoted:    `public."Table1"`,
-			Unquoted:  `public.Table1`,
-			MinQuoted: `public."Table1"`,
-		},
-		Unqualified: identifier{
-			Quoted:    `"Table1"`,
-			Unquoted:  `Table1`,
-			MinQuoted: `"Table1"`,
-		},
-		MinQualified: identifier{
-			Quoted:    `"Table1"`,
-			Unquoted:  `Table1`,
-			MinQuoted: `"Table1"`,
-		},
+	if targetSchema != "" && targetTable != "" {
+		targetName = sqlname.NewObjectName(YUGABYTEDB, targetSchema, targetSchema, targetTable)
 	}
-	assert.Equal(expectedTableName, tableName)
-	assert.Equal(tableName.MinQualified, tableName.Unqualified)
+	return NewNameTuple(reg.role, sourceName, targetName)
 }
-
-func TestPGNonDefaultSchemaCaseSensitiveTableName(t *testing.T) {
-	assert := assert.New(t)
-	// Test NewTableName() with PostgreSQL and default schema "public" and
-	// a case-sensitive name with mixed cases.
-	tableName := newObjectName(POSTGRESQL, "public", "schema1", "Table1")
-	assert.NotNil(tableName)
-	expectedTableName := &ObjectName{
-		SchemaName:        "schema1",
-		FromDefaultSchema: false,
-		Qualified: identifier{
-			Quoted:    `schema1."Table1"`,
-			Unquoted:  `schema1.Table1`,
-			MinQuoted: `schema1."Table1"`,
-		},
-		Unqualified: identifier{
-			Quoted:    `"Table1"`,
-			Unquoted:  `Table1`,
-			MinQuoted: `"Table1"`,
-		},
-		MinQualified: identifier{
-			Quoted:    `schema1."Table1"`,
-			Unquoted:  `schema1.Table1`,
-			MinQuoted: `schema1."Table1"`,
-		},
-	}
-	assert.Equal(expectedTableName, tableName)
-	assert.Equal(tableName.MinQualified, tableName.Qualified)
-}
-
-//=====================================================
-
-func TestOracleDefaultSchemaCaseInsensitiveTableName(t *testing.T) {
-	assert := assert.New(t)
-	// Test NewTableName() with Oracle and default schema "SAKILA" and
-	// a table name belonging to default schema.
-	tableName := newObjectName(ORACLE, "SAKILA", "SAKILA", "TABLE1")
-	assert.NotNil(tableName)
-	expectedTableName := &ObjectName{
-		SchemaName:        "SAKILA",
-		FromDefaultSchema: true,
-		Qualified: identifier{
-			Quoted:    `SAKILA."TABLE1"`,
-			Unquoted:  `SAKILA.TABLE1`,
-			MinQuoted: `SAKILA.TABLE1`,
-		},
-		Unqualified: identifier{
-			Quoted:    `"TABLE1"`,
-			Unquoted:  `TABLE1`,
-			MinQuoted: `TABLE1`,
-		},
-		MinQualified: identifier{
-			Quoted:    `"TABLE1"`,
-			Unquoted:  `TABLE1`,
-			MinQuoted: `TABLE1`,
-		},
-	}
-	assert.Equal(expectedTableName, tableName)
-	assert.Equal(tableName.MinQualified, tableName.Unqualified)
-}
-
-func TestOracleNonDefaultSchemaCaseInsensitiveTableName(t *testing.T) {
-	assert := assert.New(t)
-	// Test NewTableName() with Oracle and default schema "SAKILA" and
-	// a table name belonging to a non-default schema.
-	tableName := newObjectName(ORACLE, "SAKILA", "SCHEMA1", "TABLE1")
-	assert.NotNil(tableName)
-	expectedTableName := &ObjectName{
-		SchemaName:        "SCHEMA1",
-		FromDefaultSchema: false,
-		Qualified: identifier{
-			Quoted:    `SCHEMA1."TABLE1"`,
-			Unquoted:  `SCHEMA1.TABLE1`,
-			MinQuoted: `SCHEMA1.TABLE1`,
-		},
-		Unqualified: identifier{
-			Quoted:    `"TABLE1"`,
-			Unquoted:  `TABLE1`,
-			MinQuoted: `TABLE1`,
-		},
-		MinQualified: identifier{
-			Quoted:    `SCHEMA1."TABLE1"`,
-			Unquoted:  `SCHEMA1.TABLE1`,
-			MinQuoted: `SCHEMA1.TABLE1`,
-		},
-	}
-	assert.Equal(expectedTableName, tableName)
-	assert.Equal(tableName.MinQualified, tableName.Qualified)
-}
-
-func TestOracleDefaultSchemaCaseSensitiveTableName(t *testing.T) {
-	assert := assert.New(t)
-	// Test NewTableName() with Oracle and default schema "SAKILA" and
-	// a case-sensitive name with mixed cases.
-	tableName := newObjectName(ORACLE, "SAKILA", "SAKILA", "Table1")
-	assert.NotNil(tableName)
-	expectedTableName := &ObjectName{
-		SchemaName:        "SAKILA",
-		FromDefaultSchema: true,
-		Qualified: identifier{
-			Quoted:    `SAKILA."Table1"`,
-			Unquoted:  `SAKILA.Table1`,
-			MinQuoted: `SAKILA."Table1"`,
-		},
-		Unqualified: identifier{
-			Quoted:    `"Table1"`,
-			Unquoted:  `Table1`,
-			MinQuoted: `"Table1"`,
-		},
-		MinQualified: identifier{
-			Quoted:    `"Table1"`,
-			Unquoted:  `Table1`,
-			MinQuoted: `"Table1"`,
-		},
-	}
-	assert.Equal(expectedTableName, tableName)
-	assert.Equal(tableName.MinQualified, tableName.Unqualified)
-}
-
-func TestOracleNonDefaultSchemaCaseSensitiveTableName(t *testing.T) {
-	assert := assert.New(t)
-	// Test NewTableName() with Oracle and default schema "SAKILA" and
-	// a case-sensitive name with mixed cases.
-	tableName := newObjectName(ORACLE, "SAKILA", "SCHEMA1", "Table1")
-	assert.NotNil(tableName)
-	expectedTableName := &ObjectName{
-		SchemaName:        "SCHEMA1",
-		FromDefaultSchema: false,
-		Qualified: identifier{
-			Quoted:    `SCHEMA1."Table1"`,
-			Unquoted:  `SCHEMA1.Table1`,
-			MinQuoted: `SCHEMA1."Table1"`,
-		},
-		Unqualified: identifier{
-			Quoted:    `"Table1"`,
-			Unquoted:  `Table1`,
-			MinQuoted: `"Table1"`,
-		},
-		MinQualified: identifier{
-			Quoted:    `SCHEMA1."Table1"`,
-			Unquoted:  `SCHEMA1.Table1`,
-			MinQuoted: `SCHEMA1."Table1"`,
-		},
-	}
-	assert.Equal(expectedTableName, tableName)
-	assert.Equal(tableName.MinQualified, tableName.Qualified)
-}
-
-//=====================================================
-
-func TestMySQLDefaultSchemaCaseSensitiveLowerCaseTableName(t *testing.T) {
-	assert := assert.New(t)
-	// Test NewTableName() with MySQL and default schema "sakila" and
-	// a table name belonging to default schema.
-	tableName := newObjectName(MYSQL, "sakila", "sakila", "table1")
-	assert.NotNil(tableName)
-	expectedTableName := &ObjectName{
-		SchemaName:        "sakila",
-		FromDefaultSchema: true,
-		Qualified: identifier{
-			Quoted:    `sakila.table1`,
-			Unquoted:  `sakila.table1`,
-			MinQuoted: `sakila.table1`,
-		},
-		Unqualified: identifier{
-			Quoted:    `table1`,
-			Unquoted:  `table1`,
-			MinQuoted: `table1`,
-		},
-		MinQualified: identifier{
-			Quoted:    `table1`,
-			Unquoted:  `table1`,
-			MinQuoted: `table1`,
-		},
-	}
-	assert.Equal(expectedTableName, tableName)
-	assert.Equal(tableName.MinQualified, tableName.Unqualified)
-}
-
-func TestMySQLNonDefaultSchemaCaseSensitiveLowerCaseTableName(t *testing.T) {
-	assert := assert.New(t)
-	// Test NewTableName() with MySQL and default schema "sakila" and
-	// a table name belonging to a non-default schema.
-	tableName := newObjectName(MYSQL, "sakila", "schema1", "table1")
-	assert.NotNil(tableName)
-	expectedTableName := &ObjectName{
-		SchemaName:        "schema1",
-		FromDefaultSchema: false,
-		Qualified: identifier{
-			Quoted:    `schema1.table1`,
-			Unquoted:  `schema1.table1`,
-			MinQuoted: `schema1.table1`,
-		},
-		Unqualified: identifier{
-			Quoted:    `table1`,
-			Unquoted:  `table1`,
-			MinQuoted: `table1`,
-		},
-		MinQualified: identifier{
-			Quoted:    `schema1.table1`,
-			Unquoted:  `schema1.table1`,
-			MinQuoted: `schema1.table1`,
-		},
-	}
-	assert.Equal(expectedTableName, tableName)
-	assert.Equal(tableName.MinQualified, tableName.Qualified)
-}
-
-func TestMySQLDefaultSchemaCaseSensitiveMixedCaseTableName(t *testing.T) {
-	assert := assert.New(t)
-	// Test NewTableName() with MySQL and default schema "sakila" and
-	// a case-sensitive name with mixed cases.
-	tableName := newObjectName(MYSQL, "sakila", "sakila", "Table1")
-	assert.NotNil(tableName)
-	expectedTableName := &ObjectName{
-		SchemaName:        "sakila",
-		FromDefaultSchema: true,
-		Qualified: identifier{
-			Quoted:    `sakila.Table1`,
-			Unquoted:  `sakila.Table1`,
-			MinQuoted: `sakila.Table1`,
-		},
-		Unqualified: identifier{
-			Quoted:    `Table1`,
-			Unquoted:  `Table1`,
-			MinQuoted: `Table1`,
-		},
-		MinQualified: identifier{
-			Quoted:    `Table1`,
-			Unquoted:  `Table1`,
-			MinQuoted: `Table1`,
-		},
-	}
-	assert.Equal(expectedTableName, tableName)
-	assert.Equal(tableName.MinQualified, tableName.Unqualified)
-}
-
-func TestMySQLNonDefaultSchemaCaseSensitiveUpperCaseTableName(t *testing.T) {
-	assert := assert.New(t)
-	// Test NewTableName() with MySQL and default schema "sakila" and
-	// a case-sensitive name with all upper case letters.
-	tableName := newObjectName(MYSQL, "sakila", "schema1", "TABLE1")
-	assert.NotNil(tableName)
-	expectedTableName := &ObjectName{
-		SchemaName:        "schema1",
-		FromDefaultSchema: false,
-		Qualified: identifier{
-			Quoted:    `schema1.TABLE1`,
-			Unquoted:  `schema1.TABLE1`,
-			MinQuoted: `schema1.TABLE1`,
-		},
-		Unqualified: identifier{
-			Quoted:    `TABLE1`,
-			Unquoted:  `TABLE1`,
-			MinQuoted: `TABLE1`,
-		},
-		MinQualified: identifier{
-			Quoted:    `schema1.TABLE1`,
-			Unquoted:  `schema1.TABLE1`,
-			MinQuoted: `schema1.TABLE1`,
-		},
-	}
-	assert.Equal(expectedTableName, tableName)
-	assert.Equal(tableName.MinQualified, tableName.Qualified)
-}
-
-//=====================================================
 
 func TestNameTuple(t *testing.T) {
 	assert := assert.New(t)
+	sourceName := sqlname.NewObjectName(ORACLE, "SAKILA", "SAKILA", "TABLE1")
+	targetName := sqlname.NewObjectName(YUGABYTEDB, "public", "public", "table1")
 
-	ntup := &NameTuple{
-		SourceName: newObjectName(ORACLE, "SAKILA", "SAKILA", "TABLE1"),
-		TargetName: newObjectName(YUGABYTEDB, "public", "public", "table1"),
-	}
-	ntup.SetMode(TARGET_DB_IMPORTER_ROLE)
+	ntup := NewNameTuple(TARGET_DB_IMPORTER_ROLE, sourceName, targetName)
+
 	assert.Equal(ntup.CurrentName, ntup.TargetName)
 	assert.Equal(ntup.ForUserQuery(), `public."table1"`)
 	schemaName, tableName := ntup.ForCatalogQuery()
 	assert.Equal(schemaName, `public`)
 	assert.Equal(tableName, `table1`)
 
-	ntup.SetMode(SOURCE_REPLICA_DB_IMPORTER_ROLE)
+	ntup = NewNameTuple(SOURCE_REPLICA_DB_IMPORTER_ROLE, sourceName, targetName)
+
 	assert.Equal(ntup.CurrentName, ntup.SourceName)
 	assert.Equal(ntup.ForUserQuery(), `SAKILA."TABLE1"`)
 	schemaName, tableName = ntup.ForCatalogQuery()
 	assert.Equal(schemaName, `SAKILA`)
 	assert.Equal(tableName, `TABLE1`)
 
-	ntup.SetMode(SOURCE_DB_EXPORTER_ROLE)
+	ntup = NewNameTuple(SOURCE_DB_EXPORTER_ROLE, sourceName, targetName)
 	assert.Equal(ntup.CurrentName, ntup.SourceName)
 
-	ntup.SetMode(TARGET_DB_EXPORTER_FF_ROLE)
+	ntup = NewNameTuple(TARGET_DB_EXPORTER_FF_ROLE, sourceName, targetName)
 	assert.Equal(ntup.CurrentName, ntup.TargetName)
 }
 
 func TestNameTupleMatchesPattern(t *testing.T) {
 	assert := assert.New(t)
-
-	ntup := &NameTuple{
-		SourceName: newObjectName(ORACLE, "SAKILA", "SAKILA", "TABLE1"),
-		TargetName: newObjectName(YUGABYTEDB, "public", "sakila", "table1"),
-	}
-	ntup.SetMode(TARGET_DB_IMPORTER_ROLE)
+	sourceName := sqlname.NewObjectName(ORACLE, "SAKILA", "SAKILA", "TABLE1")
+	targetName := sqlname.NewObjectName(YUGABYTEDB, "public", "sakila", "table1")
+	ntup := NewNameTuple(TARGET_DB_IMPORTER_ROLE, sourceName, targetName)
 
 	testCases := []struct {
 		pattern string
@@ -434,37 +105,6 @@ func TestNameTupleMatchesPattern(t *testing.T) {
 	}
 }
 
-//=====================================================
-
-var oracleToYBNameRegistry = &NameRegistry{
-	SourceDBType:              ORACLE,
-	role:                      TARGET_DB_IMPORTER_ROLE,
-	SourceDBSchemaNames:       []string{"SAKILA"},
-	YBSchemaNames:             []string{"public"},
-	DefaultSourceDBSchemaName: "SAKILA",
-	DefaultYBSchemaName:       "public",
-	//DefaultSourceReplicaDBSchemaName: "SAKILA_FF", // Will be set using SetDefaultSourceReplicaDBSchemaName().
-	SourceDBTableNames: map[string][]string{
-		"SAKILA": {`TABLE1`, `TABLE2`, `Table2`, `MixedCaps`, `MixedCaps1`, `MixedCAPS1`, `lower_caps`},
-	},
-	YBTableNames: map[string][]string{
-		"public": {"table1", "table2", `Table2`, `mixedcaps`, `MixedCaps1`, `MixedCAPS1`, "lower_caps"},
-	},
-}
-
-func buildNameTuple(reg *NameRegistry, sourceSchema, sourceTable, targetSchema, targetTable string) *NameTuple {
-	ntup := &NameTuple{}
-
-	if sourceSchema != "" && sourceTable != "" {
-		ntup.SourceName = newObjectName(reg.SourceDBType, sourceSchema, sourceSchema, sourceTable)
-	}
-	if targetSchema != "" && targetTable != "" {
-		ntup.TargetName = newObjectName(YUGABYTEDB, targetSchema, targetSchema, targetTable)
-	}
-	ntup.SetMode(reg.role)
-	return ntup
-}
-
 func TestNameRegistrySuccessfulLookup(t *testing.T) {
 	assert := assert.New(t)
 	require := require.New(t)
@@ -477,7 +117,7 @@ func TestNameRegistrySuccessfulLookup(t *testing.T) {
 
 	var testCases = []struct {
 		tableNames []string
-		expected   *NameTuple
+		expected   *sqlname.NameTuple
 	}{
 		{[]string{
 			// YB side variants:
@@ -582,7 +222,7 @@ func TestDifferentSchemaInSameDBAsSourceReplica1(t *testing.T) {
 
 	var testCases = []struct {
 		tableNames []string
-		expected   *NameTuple
+		expected   *sqlname.NameTuple
 	}{
 		{[]string{
 			// YB side variants:
